requirements.txt
requests==2.31.0
elevenlabs==1.9.0
Pillow==10.3.0
python-dotenv==1.0.1
